requirements.txt
Flask==3.0.0 
Flask-CORS==4.0.0 
firebase-admin==6.2.0 
cryptography==41.0.7 
python-dotenv==1.0.0 
Werkzeug==3.0.1 
gunicorn==21.2.0 
